mmai-hackathon-main/pyproject.toml
#############################################
# Build system                             #
#############################################
[build-system]
requires = ["setuptools>=66", "wheel"]
build-backend = "setuptools.build_meta"

#############################################
# Project metadata                          #
#############################################
[project]
name = "mmai-hackathon"
version = "0.1.0"
description = "Base source code for MMAI workshop Hackathon."
readme = "README.md"
requires-python = ">=3.10, <3.13"
dependencies = [
    "numpy>=2.0.0",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "pydicom>=3.0.1",
    "rdkit>=2025.3.6",
    "scikit-learn>=1.6.1",
    "torch>=2.6.0",
    "torch-geometric>=2.6.0",
    "wfdb>=4.3.0",
]
authors = [
    { name = "Shuo Zhou", email = "shuo.zhou@sheffield.ac.uk" },
    { name = "Xianyuan Liu", email = "xianyuan.liu@sheffield.ac.uk" },
    { name = "Wenrui Fan", email = "wenrui.fan@sheffield.ac.uk" },
    { name = "Mohammod N. I. Suvon", email = "m.suvon@sheffield.ac.uk" },
    { name = "L. M. Riza Rizky", email = "l.m.rizky@sheffield.ac.uk" },
]

[project.optional-dependencies]
dev = [
    "black>=25.1.0",
    "dropbox>=12.0.2",
    "isort>=6.0.1",
    "flake8>=7.0.0",
    "mypy>=1.10.0",
    "pre-commit>=4.3.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
]

#############################################
# Tooling configuration                     #
#############################################
[tool.black]
# https://github.com/psf/black
line-length = 120
target-version = ["py311"]

[tool.isort]
known_first_party = ["mmai25_hackathon", "tests"]
profile = "black"
line_length = 120
force_sort_within_sections = false
order_by_type = false

[tool.mypy]
disable_error_code = ["assignment", "call-overload", "attr-defined"]

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
addopts = "--cov=mmai25_hackathon --cov-report term-missing --cov-report html"

[tool.setuptools]
py-modules = ["mmai25_hackathon"]

[tool.uv.sources]
pykale = { git = "https://github.com/pykale/pykale", rev = "main" }
